Issue 3: rename folder
