+++ src/main/java/com/apuope/apuope_re/controllers/LogInController.java
@@ -1,11 +1,8 @@
 package com.apuope.apuope_re.controllers;
 
-import com.apuope.apuope_re.entities.User;
-import com.apuope.apuope_re.repositories.UserRepository;
 import com.apuope.apuope_re.services.LogInService;
-import com.apuope.apuope_re.types.UserCredentials;
+import com.apuope.apuope_re.dto.UserCredentials;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
